+++ CHANGELOG.md
@@ -3,10 +3,12 @@ Write the date in place of the "Unreleased" in the case a new version is release
 
 # Changelog
 
-## v0.1.0b9 (Unreleased)
+
+## Unreleased
 
 ### Added
 
+- Added support for explicit units in numpy datetime64 dtypes.
 - `dtype` property for sparse arrays in `COOAdapter` and `COOStructure`.
 
 ## v0.1.0b8 (2024-09-06)